requirements.txt
pyTelegramBotAPI==4.12.0
requests==2.32.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
selenium~=4.28.1
webdriver-manager~=4.0.2
python-telegram-bot~=21.10